Allinone/microservice/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flask-Cors
pytest
pytest-cov
pytest-flask
pytest-mock
requests
python-dateutil
